move editUser service and query functions to external files in modules folder and edit usersController to use those functions
